+++ TsiErp.Entities/Entities/PlanningManagement/MRPLine/Dtos/SelectMRPLinesDto.cs
@@ -52,7 +52,7 @@ namespace TsiErp.Entities.Entities.PlanningManagement.MRPLine.Dtos
         /// <summary>
         /// Miktar
         /// </summary>
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         /// <summary>
         /// Satır No
         /// </summary>
@@ -84,7 +84,7 @@ namespace TsiErp.Entities.Entities.PlanningManagement.MRPLine.Dtos
         /// <summary>
         /// İhtiyaç Miktar
         /// </summary>
-        public int RequirementAmount { get; set; }
+        public decimal RequirementAmount { get; set; }
         /// <summary>
         /// Stok Miktarı
         /// </summary>
