+++ TsiErp.Entities/Entities/PlanningManagement/MRPLine/Dtos/UpdateMRPLinesDto.cs
@@ -39,7 +39,7 @@ namespace TsiErp.Entities.Entities.PlanningManagement.MRPLine.Dtos
         /// <summary>
         /// Miktar
         /// </summary>
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         /// <summary>
         /// Satır No
         /// </summary>
@@ -59,7 +59,7 @@ namespace TsiErp.Entities.Entities.PlanningManagement.MRPLine.Dtos
         /// <summary>
         /// İhtiyaç Miktar
         /// </summary>
-        public int RequirementAmount { get; set; }
+        public decimal RequirementAmount { get; set; }
         /// <summary>
         /// Sipariş Kabul ID
         /// </summary>
